Moved formatDateTime to util and added tests
