+++ frontend/app/module/election/page/ElectionStatusPage.tsx
@@ -26,6 +26,7 @@ import {
   ProgressBarColorClass,
   Table,
 } from "@kiesraad/ui";
+import { formatDateTime } from "@kiesraad/util";
 
 import cls from "./ElectionStatusPage.module.css";
 
@@ -103,33 +104,9 @@ function getTableRowForCategory(category: StatusCategory, polling_station: Polli
     );
   }
 
-  function formatFinishedAt(date: Date) {
-    const today = new Date();
-    const timeString = date.toLocaleTimeString(t("date_locale"), { hour: "numeric", minute: "numeric" });
-    if (
-      date.getDate() === today.getDate() &&
-      date.getMonth() === today.getMonth() &&
-      date.getFullYear() === today.getFullYear()
-    ) {
-      // Today
-      return `${t("today")} ${timeString}`;
-    } else if (Math.round(Math.abs(Number(today) - Number(date)) / (24 * 60 * 60 * 1000)) < 7) {
-      // Within the past 6 days
-      return date.toLocaleString(t("date_locale"), {
-        weekday: "long",
-        hour: "numeric",
-        minute: "numeric",
-      });
-    } else {
-      // More than 6 days ago
-      const dateString = date.toLocaleDateString(t("date_locale"), { day: "numeric", month: "short" });
-      return `${dateString} ${timeString}`;
-    }
-  }
-
   const finishedAtCell = (
     <Table.Cell key={`${polling_station.id}-time`}>
-      {polling_station.finished_at ? formatFinishedAt(new Date(polling_station.finished_at * 1000)) : ""}
+      {polling_station.finished_at ? formatDateTime(new Date(polling_station.finished_at * 1000)) : ""}
     </Table.Cell>
   );
   const progressCell = (
