+++ frontend/lib/util/format.ts
@@ -1,3 +1,5 @@
+import { t } from "@kiesraad/i18n";
+
 const numberFormatter = new Intl.NumberFormat("nl-NL", {
   maximumFractionDigits: 0,
 });
@@ -26,4 +28,28 @@ export function validateNumberString(s: string | null | undefined) {
   if (s === null || s === undefined || s === "") return false;
   const result = s.trim();
   return !!result.match(/^(\d*\.?)$|^(\d{1,3}(\.\d{3})+\.?)$/g);
+}
+
+export function formatDateTime(date: Date) {
+  const today = new Date();
+  const timeString = date.toLocaleTimeString(t("date_locale"), { hour: "numeric", minute: "numeric" });
+  if (
+    date.getDate() === today.getDate() &&
+    date.getMonth() === today.getMonth() &&
+    date.getFullYear() === today.getFullYear()
+  ) {
+    // Today
+    return `${t("today")} ${timeString}`;
+  } else if (Math.round(Math.abs(Number(today) - Number(date)) / (24 * 60 * 60 * 1000)) < 7) {
+    // Within the past 6 days
+    return date.toLocaleString(t("date_locale"), {
+      weekday: "long",
+      hour: "numeric",
+      minute: "numeric",
+    });
+  } else {
+    // More than 6 days ago (or in the future)
+    const dateString = date.toLocaleDateString(t("date_locale"), { day: "numeric", month: "short" });
+    return `${dateString} ${timeString}`;
+  }
 }
